requirements.txt
anyio==4.0.0
attrs==23.1.0
Automat==22.10.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
decorator==5.1.1
exceptiongroup==1.1.3
filelock==3.12.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
loguru==0.7.1
lxml==4.9.3
nested-lookup==0.2.25
numpy==1.25.2
packaging==23.1
parsel==1.8.1
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-decouple==3.8
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
scrapfly-sdk==0.8.8
Scrapy==2.10.1
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tldextract==3.5.0
Twisted==22.10.0
typing_extensions==4.7.1
urllib3==2.0.4
w3lib==2.1.2
zope.interface==6.0
